backend/pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = ["Bernardo Meirelles <bernardomrl@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg = "^3.1.16"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
mypy_path = "src"